SMILES2MOL/pyproject.toml
[build-system]
requires = ["setuptools>=65", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "smiles2mol"
version = "0.1.0"
description = "SMILES to RDKit Mol with 3D conformer generation"
authors = [
    { name="Your Name", email="your_email@example.com" }
]
readme = "README.md"
requires-python = ">=3.10,<3.14"
license = "MIT"
dependencies = [
    "rdkit-pypi; sys_platform != 'win32'"
]

[tool.setuptools]
package-dir = {"" = "src"}
packages = ["smiles2mol"]
